HphpArray compact layout
ArrayData gets on a diet, down to 32 bytes
devirtualized destructor call in HphpArray::release()
aggressively inlined constructors for ArrayData (do away with defaulted arguments)
eliminate a dead memset() for the hashtable in HphpArray(uint size, const TypedValue* values)

+++ hphp/runtime/vm/name_value_table_wrapper.cpp
@@ -252,7 +252,7 @@ Variant NameValueTableWrapper::key() const {
   return null_variant;
 }
 
-Variant NameValueTableWrapper::value(ssize_t& pos) const {
+Variant NameValueTableWrapper::value(int32_t& pos) const {
   return current();
 }
